requirements.txt
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses==0.6
google-api-core==1.31.5
google-api-python-client==2.6.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-drive==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
packaging==24.0
protobuf==5.26.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.2.0
six==1.16.0
uritemplate==3.0.1
urllib3==2.2.1
